requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow

black
isort